+++ PostProcessing/Runtime/Utils/RuntimeUtilities.cs
@@ -1009,7 +1009,17 @@ namespace UnityEngine.Rendering.PostProcessing
                 && layer.antialiasingMode == PostProcessLayer.Antialiasing.TemporalAntialiasing
                 && layer.temporalAntialiasing.IsSupported();
         }
-
+#if UNITY_2017_3_OR_NEWER
+        /// <summary>
+        /// Checks if dynamic resolution is enabled on a given camera.
+        /// </summary>
+        /// <param name="camera">The camera to check</param>
+        /// <returns><c>true</c> if dynamic resolution is enabled, <c>false</c> otherwise</returns>
+        public static bool IsDynamicResolutionEnabled(Camera camera)
+        {
+            return camera.allowDynamicResolution || (camera.targetTexture != null && camera.targetTexture.useDynamicScale);
+        }
+#endif
         /// <summary>
         /// Gets all scene objects in the hierarchy, including inactive objects. This method is slow
         /// on large scenes and should be used with extreme caution.
@@ -1025,9 +1035,8 @@ namespace UnityEngine.Rendering.PostProcessing
             foreach (var root in roots)
             {
                 queue.Enqueue(root.transform);
-                var comp = root.GetComponent<T>();
-
-                if (comp != null)
+                
+                if (root.TryGetComponent<T>(out var comp))
                     yield return comp;
             }
 
@@ -1036,9 +1045,8 @@ namespace UnityEngine.Rendering.PostProcessing
                 foreach (Transform child in queue.Dequeue())
                 {
                     queue.Enqueue(child);
-                    var comp = child.GetComponent<T>();
-
-                    if (comp != null)
+                    
+                    if (child.TryGetComponent<T>(out var comp))
                         yield return comp;
                 }
             }