+++ CHANGELOG.md
@@ -4,10 +4,18 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2023-05-11
+
+### Fixed
+- Fixed some post effects not scaling correctly when dynamic resolution is enabled on a camera's target texture (case 1166603)
+- Fixed obsolete API call GetScriptingDefineSymbolsForGroup
+- Remove support for OpenGL ES 2.0 in Unity 2023.1 or newer
+
 ## [3.2.2] - 2022-04-05
 
 ### Fixed
 - Fixed FXAA artefact when trying to preserve alpha channel output.
+- Fixed MSVO aliasing artifact by using a z-Bias parameter. (case 1375337)
 
 ## [3.2.1] - 2022-01-12